refactor(http2): per-concern split of server/response.ts
Applies the install*Methods(proto) pattern (PRs #261, #273, #309) to
packages/node/http2/src/server/response.ts. The 903-LoC monolith is
split into four per-concern sibling modules under a new
src/server/response/ directory:

  - response/headers.ts            (159 LoC) — setHeader / getHeader /
                                               removeHeader / hasHeader /
                                               getHeaderNames / getHeaders /
                                               appendHeader / flushHeaders /
                                               writeHead / respond /
                                               writeContinue /
                                               writeEarlyHints /
                                               addTrailers / setTimeout
  - response/stream-io.ts          (168 LoC) — _startStreaming / _write /
                                               _final / _sendBatchResponse /
                                               end. The Writable lifecycle
                                               hooks are deliberately
                                               excluded from the
                                               declaration-merge interface
                                               to avoid colliding with the
                                               inherited
                                               Writable._write(chunk: any,
                                               ...) overload; they still
                                               get attached to the runtime
                                               prototype via Object.assign.
  - response/respond-with-file.ts  (271 LoC) — respondWithFD /
                                               respondWithFile + the
                                               file-local _respondFromFD /
                                               _fdPath helpers + the
                                               exported StatCheck type.
  - response/push.ts               (289 LoC) — both classes' pushStream
                                               entry points +
                                               Http2ServerResponse.
                                               createPushResponse + the
                                               file-local
                                               _makeDetachedSoupMessage
                                               factory. Push surfaces of
                                               Http2ServerResponse and
                                               ServerHttp2Stream stay
                                               together because they share
                                               the PUSH_PROMISE allocator +
                                               frame-builder coupling.

Same shape as PR #309: each sibling exports a typed *Methods interface
declaration-merged into the class via `declare module '../response.js'`
and an install<Group>Methods(proto) function copying its ThisType<>-
bound methods object onto the prototype. The base file wires all four in
after the class declarations with the deliberate side-effect-import-
next-to-named-import pattern so tsc preserves the augmentations in the
emitted .d.ts.

Base file response.ts: 903 → 288 LoC, owning the type definitions
(Http2NativeBackend, StatCheck), both class declarations (fields +
constructors + read-only getters), and ServerHttp2Stream's delegate
methods (the ones that just forward to _res.* without their own logic —
write / end / destroy / close / priority / setTimeout / sendTrailers /
additionalHeaders / respondWithFD / respondWithFile).

Previously-private fields on both classes were demoted to default
visibility with @internal JSDoc so sibling modules can access them via
this.* (matches the webgl2-rendering-context split pattern).

Public surface preserved verbatim: every export from @gjsify/http2 /
@gjsify/http2/server works unchanged. 102 Node + 231 GJS tests still
green; tsc --noEmit clean.
