+++ packages/node/http2/src/server/response.ts
@@ -5,32 +5,25 @@
 // the parent response; Response.createPushResponse → calls pushStream and
 // extracts the child's response. Splitting them across files would need
 // type-only imports both ways and dead-class hooks, with no readability
-// gain. Single file = ~700 LoC, still well below the per-file ceiling the
-// monorepo's other classes (canvas-rendering-context-2d, http2-native
-// SessionBridge.vala) live at.
+// gain. The method bodies live in per-concern siblings under
+// `./response/` and are wired into the prototypes below via the same
+// `install*Methods(proto)` pattern as the webgl2-rendering-context split
+// (PR #309) — typed `*Methods` interface declaration-merged into the
+// class plus an `install<Group>Methods(proto)` function that copies the
+// implementations onto the prototype.
 //
-// The `_respondFromFD` / `_makeDetachedSoupMessage` / `_fdPath` helpers
-// live here too because they're only ever called from Response methods
-// (`respondWithFD` / `respondWithFile` / `pushStream`).
+//   ./response/headers.ts            header management + writeHead + respond + setTimeout
+//   ./response/stream-io.ts          _write / _final / _startStreaming / end
+//   ./response/respond-with-file.ts  respondWithFD / respondWithFile + helpers
+//   ./response/push.ts               pushStream / createPushResponse + ServerHttp2Stream.pushStream
 
-import Soup from '@girs/soup-3.0';
+import type Soup from '@girs/soup-3.0';
 import { EventEmitter } from 'node:events';
 import { Writable } from 'node:stream';
 import { Buffer } from 'node:buffer';
-import { read as fsRead, statSync, openSync, closeSync, type Stats } from 'node:fs';
-import type { OutgoingHttpHeaders } from 'node:http';
 import { constants } from '../protocol.js';
 import type { ServerHttp2Session } from './session.js';
-
-/**
- * StatCheck callback signature — matches Node's
- * `http2.ServerStreamFileResponseOptions.statCheck`:
- * `(stats: fs.Stats, headers: OutgoingHttpHeaders, statOptions: { offset, length }) => void`.
- * The user mutates headers based on stat results; returning `false` cancels
- * the send (Node behaviour — wired through `_respondFromFD`).
- */
-type StatCheckOptions = { offset?: number; length?: number };
-type StatCheck = (stats: Stats, headers: OutgoingHttpHeaders, statOptions: StatCheckOptions) => void | boolean;
+import type { StatCheck } from './response/respond-with-file.js';
 
 /**
  * Per-stream backend that routes writes through `SessionBridge.submit_*`
@@ -63,14 +56,20 @@ export class Http2ServerResponse extends Writable {
     finished = false;
     sendDate = true;
 
-    private _soupMsg: Soup.ServerMessage | null;
-    private _nativeBackend: Http2NativeBackend | null;
-    private _headers: Map<string, string | string[]> = new Map();
-    private _streaming = false;
-    private _timeoutTimer: ReturnType<typeof setTimeout> | null = null;
-    private _stream: ServerHttp2Stream | null = null;
-    /** Detached responses (PUSH_PROMISE children) buffer their output. */
-    private _detachedBody: Buffer[] | null = null;
+    /** @internal — sibling response/*.ts modules access this directly. */
+    _soupMsg: Soup.ServerMessage | null;
+    /** @internal */
+    _nativeBackend: Http2NativeBackend | null;
+    /** @internal */
+    _headers: Map<string, string | string[]> = new Map();
+    /** @internal */
+    _streaming = false;
+    /** @internal */
+    _timeoutTimer: ReturnType<typeof setTimeout> | null = null;
+    /** @internal */
+    _stream: ServerHttp2Stream | null = null;
+    /** @internal Detached responses (PUSH_PROMISE children) buffer their output. */
+    _detachedBody: Buffer[] | null = null;
 
     get stream(): ServerHttp2Stream | null {
         return this._stream;
@@ -111,342 +110,6 @@ export class Http2ServerResponse extends Writable {
         this._nativeBackend = nativeBackend;
         if (soupMsg === null && nativeBackend === null) this._detachedBody = [];
     }
-
-    setHeader(name: string, value: string | number | string[]): this {
-        this._headers.set(name.toLowerCase(), typeof value === 'number' ? String(value) : value);
-        return this;
-    }
-
-    getHeader(name: string): string | string[] | undefined {
-        return this._headers.get(name.toLowerCase());
-    }
-
-    removeHeader(name: string): void {
-        this._headers.delete(name.toLowerCase());
-    }
-
-    hasHeader(name: string): boolean {
-        return this._headers.has(name.toLowerCase());
-    }
-
-    getHeaderNames(): string[] {
-        return Array.from(this._headers.keys());
-    }
-
-    getHeaders(): Record<string, string | string[]> {
-        const result: Record<string, string | string[]> = {};
-        for (const [key, value] of this._headers) {
-            result[key] = value;
-        }
-        return result;
-    }
-
-    appendHeader(name: string, value: string | string[]): this {
-        const lower = name.toLowerCase();
-        const existing = this._headers.get(lower);
-        if (existing === undefined) {
-            this._headers.set(lower, value);
-        } else if (Array.isArray(existing)) {
-            if (Array.isArray(value)) existing.push(...value);
-            else existing.push(value);
-        } else {
-            this._headers.set(
-                lower,
-                Array.isArray(value) ? [existing as string, ...value] : [existing as string, value],
-            );
-        }
-        return this;
-    }
-
-    flushHeaders(): void {
-        if (!this.headersSent) this.headersSent = true;
-    }
-
-    writeHead(
-        statusCode: number,
-        statusMessage?: string | Record<string, string | string[]>,
-        headers?: Record<string, string | string[]>,
-    ): this {
-        this.statusCode = statusCode;
-        if (typeof statusMessage === 'object') {
-            headers = statusMessage;
-            statusMessage = undefined;
-        }
-        if (typeof statusMessage === 'string') this.statusMessage = statusMessage;
-        if (headers) {
-            for (const [key, value] of Object.entries(headers)) {
-                this.setHeader(key, value);
-            }
-        }
-        return this;
-    }
-
-    // http2 session-API alias — extracts :status from headers map
-    respond(headers: Record<string, string | string[] | number>, options?: { endStream?: boolean }): void {
-        const status = Number(headers[':status'] ?? 200);
-        const rest: Record<string, string | string[]> = {};
-        for (const [k, v] of Object.entries(headers)) {
-            if (k === ':status') continue;
-            rest[k] = typeof v === 'number' ? String(v) : v;
-        }
-        this.writeHead(status, rest);
-        if (options?.endStream) this.end();
-    }
-
-    writeContinue(callback?: () => void): void {
-        if (callback) Promise.resolve().then(callback);
-    }
-
-    writeEarlyHints(_hints: Record<string, string | string[]>, callback?: () => void): void {
-        if (callback) Promise.resolve().then(callback);
-    }
-
-    addTrailers(_headers: Record<string, string>): void {}
-
-    setTimeout(msecs: number, callback?: () => void): this {
-        if (this._timeoutTimer) {
-            clearTimeout(this._timeoutTimer);
-            this._timeoutTimer = null;
-        }
-        if (callback) this.once('timeout', callback);
-        if (msecs > 0) {
-            this._timeoutTimer = setTimeout(() => {
-                this._timeoutTimer = null;
-                this.emit('timeout');
-            }, msecs);
-        }
-        return this;
-    }
-
-    private _startStreaming(): void {
-        if (this._streaming) return;
-        this._streaming = true;
-        this.headersSent = true;
-
-        if (this._timeoutTimer) {
-            clearTimeout(this._timeoutTimer);
-            this._timeoutTimer = null;
-        }
-
-        if (this._nativeBackend) {
-            // Native dispatcher: submit response headers; END_STREAM goes with the
-            // first DATA frame in _write or _final.
-            this._nativeBackend.submitResponse(this.statusCode, this.statusMessage, this._headers, false);
-            return;
-        }
-
-        if (!this._soupMsg) return; // detached push response — no Soup wire
-
-        this._soupMsg.set_status(this.statusCode, this.statusMessage || null);
-        const responseHeaders = this._soupMsg.get_response_headers();
-
-        if (this._headers.has('content-length')) {
-            responseHeaders.set_encoding(Soup.Encoding.CONTENT_LENGTH);
-        } else {
-            responseHeaders.set_encoding(Soup.Encoding.CHUNKED);
-        }
-
-        for (const [key, value] of this._headers) {
-            if (Array.isArray(value)) {
-                for (const v of value) responseHeaders.append(key, v);
-            } else {
-                responseHeaders.replace(key, value as string);
-            }
-        }
-    }
-
-    _write(chunk: string | Buffer | Uint8Array, encoding: string, callback: (error?: Error | null) => void): void {
-        const buf = Buffer.isBuffer(chunk)
-            ? chunk
-            : typeof chunk === 'string'
-              ? Buffer.from(chunk, encoding as BufferEncoding)
-              : Buffer.from(chunk);
-        this._startStreaming();
-        if (this._nativeBackend) {
-            this._nativeBackend.submitData(buf, false);
-        } else if (this._soupMsg) {
-            const responseBody = this._soupMsg.get_response_body();
-            responseBody.append(new Uint8Array(buf.buffer, buf.byteOffset, buf.byteLength));
-            this._soupMsg.unpause();
-        } else if (this._detachedBody) {
-            this._detachedBody.push(buf);
-        }
-        callback();
-    }
-
-    _final(callback: (error?: Error | null) => void): void {
-        if (this._streaming) {
-            if (this._nativeBackend) {
-                this._nativeBackend.submitData(Buffer.alloc(0), true);
-            } else if (this._soupMsg) {
-                const responseBody = this._soupMsg.get_response_body();
-                responseBody.complete();
-                this._soupMsg.unpause();
-            }
-        } else {
-            this._sendBatchResponse();
-        }
-        this.finished = true;
-        callback();
-    }
-
-    private _sendBatchResponse(): void {
-        if (this.headersSent) return;
-        this.headersSent = true;
-
-        if (this._timeoutTimer) {
-            clearTimeout(this._timeoutTimer);
-            this._timeoutTimer = null;
-        }
-
-        if (this._nativeBackend) {
-            // No body — submit headers with END_STREAM.
-            this._nativeBackend.submitResponse(this.statusCode, this.statusMessage, this._headers, true);
-            return;
-        }
-
-        if (!this._soupMsg) return;
-
-        this._soupMsg.set_status(this.statusCode, this.statusMessage || null);
-        const responseHeaders = this._soupMsg.get_response_headers();
-
-        for (const [key, value] of this._headers) {
-            if (Array.isArray(value)) {
-                for (const v of value) responseHeaders.append(key, v);
-            } else {
-                responseHeaders.replace(key, value as string);
-            }
-        }
-
-        const contentType = (this._headers.get('content-type') as string) || 'text/plain';
-        this._soupMsg.set_response(contentType, Soup.MemoryUse.COPY, new Uint8Array(0));
-    }
-
-    end(chunk?: unknown, encoding?: BufferEncoding | (() => void), callback?: () => void): this {
-        if (typeof chunk === 'function') {
-            callback = chunk as () => void;
-            chunk = undefined;
-        } else if (typeof encoding === 'function') {
-            callback = encoding;
-            encoding = undefined;
-        }
-        if (chunk != null) {
-            this.write(chunk as string | Buffer, encoding as BufferEncoding);
-        }
-        super.end(callback);
-        return this;
-    }
-
-    /**
-     * respondWithFD — stream the contents of an open file descriptor as the
-     * response body. Headers are sent once `statCheck()` (if provided) has
-     * had a chance to mutate them; payload is read in 64 KiB chunks via
-     * `fs.read()` and dispatched through the existing Soup chunked-write path.
-     *
-     * Reference: Node.js doc/api/http2.md § respondWithFD()
-     */
-    respondWithFD(
-        fd: number | { fd: number },
-        headers?: Record<string, string | string[] | number>,
-        options?: { offset?: number; length?: number; statCheck?: StatCheck },
-    ): void {
-        _respondFromFD(this, fd, headers, options ?? {}, /* closeFd */ false);
-    }
-
-    /**
-     * respondWithFile — stream a regular file by path. Opens the file with
-     * fs.openSync, runs the optional `statCheck()` callback so the user can
-     * mutate headers based on stat results (last-modified, size, etag, …),
-     * then delegates to the same FD-streaming path as `respondWithFD()`.
-     *
-     * Reference: Node.js doc/api/http2.md § respondWithFile()
-     */
-    respondWithFile(
-        path: string,
-        headers?: Record<string, string | string[] | number>,
-        options?: {
-            offset?: number;
-            length?: number;
-            statCheck?: StatCheck;
-            onError?: (err: Error) => void;
-        },
-    ): void {
-        let fd: number;
-        try {
-            fd = openSync(path, 'r');
-        } catch (err) {
-            if (options?.onError) {
-                options.onError(err as Error);
-                return;
-            }
-            throw err;
-        }
-        _respondFromFD(this, fd, headers, options ?? {}, /* closeFd */ true);
-    }
-
-    /**
-     * pushStream — request the server to push an additional resource on a
-     * fresh server-initiated stream. The Vala/nghttp2 bridge allocates the
-     * promised even stream-id and constructs the PUSH_PROMISE frame; wire-level
-     * delivery requires raw nghttp2-on-socket access that Soup does not expose,
-     * so the byte-frame is currently a no-op on the wire — but the bridge
-     * allocator and frame builder are exercised end-to-end and the callback
-     * receives a fully-usable `ServerHttp2Stream` whose `respond()` / `end()`
-     * calls write into a synthetic in-memory stream observable from tests.
-     *
-     * See STATUS.md "Open TODOs" → "http2 PUSH_PROMISE wire delivery".
-     */
-    pushStream(
-        headers: Record<string, string | string[] | number>,
-        options:
-            | { parent?: number; weight?: number; exclusive?: boolean }
-            | ((err: Error | null, pushStream: ServerHttp2Stream, headers: Record<string, string | string[]>) => void),
-        callback?: (
-            err: Error | null,
-            pushStream: ServerHttp2Stream,
-            headers: Record<string, string | string[]>,
-        ) => void,
-    ): void {
-        if (typeof options === 'function') {
-            callback = options;
-            options = {};
-        }
-        if (!callback) {
-            // Match Node behaviour: missing callback raises ERR_INVALID_ARG_TYPE
-            throw new TypeError('callback must be a function');
-        }
-        if (!this._stream) {
-            callback(new Error('No associated stream'), null as unknown as ServerHttp2Stream, {});
-            return;
-        }
-        this._stream.pushStream(headers, options, callback);
-    }
-
-    /**
-     * createPushResponse — alternate API: create a child Http2ServerResponse
-     * for the push without needing to bridge through ServerHttp2Stream. The
-     * created response shares the parent's stream allocator + bridge.
-     *
-     * Reference: Node.js doc/api/http2.md § Http2ServerResponse#createPushResponse()
-     */
-    createPushResponse(
-        headers: Record<string, string | string[] | number>,
-        callback: (err: Error | null, res: Http2ServerResponse) => void,
-    ): void {
-        if (typeof callback !== 'function') {
-            throw new TypeError('callback must be a function');
-        }
-        this.pushStream(headers, {}, (err, pushStream) => {
-            if (err) {
-                callback(err, null as unknown as Http2ServerResponse);
-                return;
-            }
-            // The synthetic ServerHttp2Stream owns its own Http2ServerResponse
-            // (created in ServerHttp2Stream.pushStream below) — extract it.
-            const res = (pushStream as unknown as { _res?: Http2ServerResponse })._res;
-            callback(null, res ?? (null as unknown as Http2ServerResponse));
-        });
-    }
 }
 
 // ─── ServerHttp2Stream ────────────────────────────────────────────────────────
@@ -458,15 +121,18 @@ export class ServerHttp2Stream extends EventEmitter {
     readonly pushAllowed: boolean;
     readonly sentHeaders: Record<string, string | string[]> = {};
 
-    private _res: Http2ServerResponse;
-    private _session: ServerHttp2Session | null;
-    private _isPushedStream: boolean;
-    /** Children pushed off this request stream (parent → array). */
-    private _pushedChildren: ServerHttp2Stream[] = [];
-    /** Cached PUSH_PROMISE frame bytes for inspection in tests. */
-    private _pushPromiseFrame: Uint8Array | null = null;
-    /** Push request headers (`:method`, `:path`, …). */
-    private _pushRequestHeaders: Record<string, string | string[]> | null = null;
+    /** @internal — sibling response/*.ts modules access this directly. */
+    _res: Http2ServerResponse;
+    /** @internal */
+    _session: ServerHttp2Session | null;
+    /** @internal */
+    _isPushedStream: boolean;
+    /** @internal Children pushed off this request stream (parent → array). */
+    _pushedChildren: ServerHttp2Stream[] = [];
+    /** @internal Cached PUSH_PROMISE frame bytes for inspection in tests. */
+    _pushPromiseFrame: Uint8Array | null = null;
+    /** @internal Push request headers (`:method`, `:path`, …). */
+    _pushRequestHeaders: Record<string, string | string[]> | null = null;
 
     get session(): ServerHttp2Session | null {
         return this._session;
@@ -598,306 +264,25 @@ export class ServerHttp2Stream extends EventEmitter {
     ): void {
         this._res.respondWithFile(path, headers, options);
     }
-
-    /**
-     * pushStream — see {@link Http2ServerResponse.pushStream} for the full
-     * contract. This is the lower-level entry point: it allocates a promised
-     * stream-id from the session-bound `GjsifyHttp2.StreamIdAllocator`, builds
-     * the PUSH_PROMISE frame via `GjsifyHttp2.FrameEncoder`, then synthesises
-     * a child `ServerHttp2Stream` whose response surface is independent of
-     * the parent's underlying SoupServerMessage.
-     */
-    pushStream(
-        headers: Record<string, string | string[] | number>,
-        options:
-            | { parent?: number; weight?: number; exclusive?: boolean }
-            | ((err: Error | null, pushStream: ServerHttp2Stream, headers: Record<string, string | string[]>) => void),
-        callback?: (
-            err: Error | null,
-            pushStream: ServerHttp2Stream,
-            headers: Record<string, string | string[]>,
-        ) => void,
-    ): void {
-        if (typeof options === 'function') {
-            callback = options;
-            options = {};
-        }
-        if (!callback) {
-            throw new TypeError('callback must be a function');
-        }
-
-        // Per RFC 7540 §8.2: pushed streams MUST NOT initiate further pushes.
-        // Node surfaces this as ERR_HTTP2_NESTED_PUSH.
-        if (this._isPushedStream) {
-            const err = Object.assign(new Error('Cannot initiate nested push streams'), {
-                code: 'ERR_HTTP2_NESTED_PUSH',
-            });
-            callback(err, null as unknown as ServerHttp2Stream, {});
-            return;
-        }
-
-        // Session-level enable_push must be honoured. Soup-backed sessions
-        // default to allowing it (we simulate the API), but a goaway/SETTINGS
-        // toggle disables further pushes.
-        if (this._session && this._session.canPush === false) {
-            const err = Object.assign(new Error('HTTP/2 server push has been disabled'), {
-                code: 'ERR_HTTP2_PUSH_DISABLED',
-            });
-            callback(err, null as unknown as ServerHttp2Stream, {});
-            return;
-        }
-
-        // Allocate the promised stream-id and build the PUSH_PROMISE frame
-        // bytes. Both go through the @gjsify/http2-native bridge when the
-        // typelib is loadable; otherwise we fall back to in-process counters.
-        let promisedId: number;
-        let frameBytes: Uint8Array | null = null;
-        let pushHeaders: Record<string, string | string[]> = {};
-
-        // Normalise pseudo-headers — Node fills in :scheme/:authority from
-        // the parent if omitted (matches refs/node/lib/internal/http2/util.js).
-        const normalised: Record<string, string | string[]> = {};
-        for (const [k, v] of Object.entries(headers)) {
-            normalised[k] = typeof v === 'number' ? String(v) : v;
-        }
-        if (!normalised[':method']) normalised[':method'] = 'GET';
-        pushHeaders = normalised;
-
-        // Native-dispatcher path: if this stream's response is backed by a
-        // native backend, submit the PUSH_PROMISE through the wire and build
-        // a pushed Http2ServerResponse whose writes route through the child
-        // backend the dispatcher just allocated. Both the promised stream id
-        // AND the PUSH_PROMISE frame come from nghttp2 itself.
-        const parentBackend = this._res.nativeBackend;
-        let pushChildBackend: Http2NativeBackend | null = null;
-        if (parentBackend) {
-            pushChildBackend = parentBackend.pushPromise(normalised);
-            if (!pushChildBackend) {
-                const err = Object.assign(new Error('No available stream ids'), {
-                    code: 'ERR_HTTP2_OUT_OF_STREAMS',
-                });
-                callback(err, null as unknown as ServerHttp2Stream, {});
-                return;
-            }
-            promisedId = pushChildBackend.streamId;
-            // No frameBytes — the bytes are already on the wire; expose `null`
-            // through `pushPromiseFrame` so consumers can distinguish the two
-            // paths during the transition window.
-            frameBytes = null;
-        } else if (this._session) {
-            promisedId = this._session._allocatePushId();
-            if (promisedId === 0) {
-                const err = Object.assign(new Error('No available stream ids'), {
-                    code: 'ERR_HTTP2_OUT_OF_STREAMS',
-                });
-                callback(err, null as unknown as ServerHttp2Stream, {});
-                return;
-            }
-            frameBytes = this._session._buildPushPromise(this.id, promisedId, normalised);
-        } else {
-            // No session attached — synthesise a counter so tests see a stable id.
-            promisedId = 2;
-        }
-
-        // Build the synthetic response surface. Native path: wire writes via
-        // the child backend. Soup path: writes land in a detached buffer
-        // reachable from `pushStream._res.detachedBody` (test inspection only,
-        // since Soup multiplexes HTTP/2 streams internally and refuses
-        // external injection).
-        const pushRes = pushChildBackend
-            ? new Http2ServerResponse(null, pushChildBackend)
-            : new Http2ServerResponse(_makeDetachedSoupMessage());
-        const pushStream = new ServerHttp2Stream(pushRes, this._session, {
-            isPushedStream: true,
-            streamId: promisedId,
-        });
-        pushStream._pushPromiseFrame = frameBytes;
-        pushStream._pushRequestHeaders = normalised;
-        pushRes._setStream(pushStream);
-        this._pushedChildren.push(pushStream);
-
-        // Match Node's contract: callback runs asynchronously after the
-        // pushStream is wired up.
-        Promise.resolve().then(() => {
-            callback!(null, pushStream, pushHeaders);
-        });
-    }
 }
 
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-// Internal-only — only called from Response/Stream methods above.
+// Wire focused method groups into the two class prototypes, same pattern as
+// the webgl2-rendering-context split (PR #309). Each module declares a
+// `*Methods` interface merged into the class via `declare module './response.js'`
+// and exposes an `install<Group>Methods(proto)` function that copies its
+// `ThisType<>`-bound method object onto the prototype. The side-effect import
+// is kept separate from the named import so tsc preserves the augmentation in
+// the emitted `.d.ts` (downstream consumers see the methods on the public type).
+import './response/headers.js';
+import { installHeaderMethods } from './response/headers.js';
+import './response/stream-io.js';
+import { installStreamIoMethods } from './response/stream-io.js';
+import './response/respond-with-file.js';
+import { installRespondWithFileMethods } from './response/respond-with-file.js';
+import './response/push.js';
+import { installPushMethods } from './response/push.js';
 
-/**
- * _makeDetachedSoupMessage — placeholder factory for push-stream Http2ServerResponse.
- *
- * Push streams have no associated SoupServerMessage (the Soup connection
- * multiplexer multiplexes them internally and refuses external injection),
- * so we hand the response a `null` Soup message and let it route writes
- * into a buffered backing store via `Http2ServerResponse._detachedBody`.
- *
- * Kept as a function (not an inline `null`) so future revisions can return
- * a real shadow message once Soup exposes the underlying nghttp2 session
- * — call sites won't have to change.
- */
-function _makeDetachedSoupMessage(): Soup.ServerMessage | null {
-    return null;
-}
-
-/**
- * _respondFromFD — common implementation behind respondWithFD / respondWithFile.
- *
- * Flow:
- *  1) statSync on the FD so the user-supplied `statCheck()` callback can
- *     mutate headers based on size / mtime / ino (Node parity).
- *  2) flushHeaders via writeHead — kicks the Soup chunked-write path.
- *  3) Read the FD in 64 KiB chunks via fs.read; pipe each chunk through
- *     `res.write()` so existing Soup pause/unpause back-pressure applies.
- *  4) On EOF, call `res.end()` and close the FD if we opened it.
- *
- * This deliberately uses `node:fs` (the gjsify polyfill) instead of
- * `Gio.UnixInputStream` so the same code path works on Node test runs.
- */
-function _respondFromFD(
-    res: Http2ServerResponse,
-    fdOrHandle: number | { fd: number },
-    headers: Record<string, string | string[] | number> | undefined,
-    options: {
-        offset?: number;
-        length?: number;
-        statCheck?: StatCheck;
-        onError?: (err: Error) => void;
-    },
-    closeFd: boolean,
-): void {
-    // Both raw numeric fds and `@gjsify/fs` FileHandle wrappers (which carry
-    // the numeric fd on `.fd`) are accepted — `fs.openSync()` returns the
-    // wrapper on GJS, a raw integer on Node.
-    const fd: number = typeof fdOrHandle === 'number' ? fdOrHandle : (fdOrHandle as { fd: number }).fd;
-    // Always hand `fs.read` / `fs.close` the numeric fd. On GJS the @gjsify/fs
-    // FileHandle wrapper registers itself under the numeric fd in its FD
-    // table — passing the wrapper object itself fails the lookup
-    // (object → "[object Object]" string key).
-    const fdArg: number = fd;
-    const finalHeaders: Record<string, string | string[] | number> = { ...headers };
-
-    // statCheck — mirrors Node's contract: lets the app mutate headers based
-    // on stat results without hand-writing fstat boilerplate.
-    if (options.statCheck) {
-        try {
-            const stat = statSync(_fdPath(fd) ?? '/proc/self/fd/' + fd);
-            // `finalHeaders` is the http2-style outgoing-headers shape
-            // (`Record<string, string | string[] | number>`) which is the same
-            // record-shape as `OutgoingHttpHeaders`; the explicit cast bridges
-            // the nominal mismatch without weakening either declaration.
-            const cont = options.statCheck(stat, finalHeaders as OutgoingHttpHeaders, options);
-            if (cont === false) {
-                if (closeFd) closeSync(fd);
-                res.end();
-                return;
-            }
-        } catch (err) {
-            if (options.onError) {
-                options.onError(err as Error);
-                if (closeFd) closeSync(fd);
-                return;
-            }
-            // Continue without statCheck — Node's behaviour is to skip silently
-            // when fstat fails (the FD will fail later in the read loop anyway).
-        }
-    }
-
-    // Headers go out first.
-    const status = Number(finalHeaders[':status'] ?? 200);
-    delete finalHeaders[':status'];
-    const sanitised: Record<string, string | string[]> = {};
-    for (const [k, v] of Object.entries(finalHeaders)) {
-        sanitised[k] = typeof v === 'number' ? String(v) : v;
-    }
-    res.writeHead(status, sanitised);
-    res.flushHeaders();
-
-    const startOffset = Math.max(0, options.offset ?? 0);
-    const totalLength = options.length;
-    const CHUNK = 64 * 1024;
-    const buffer = Buffer.alloc(CHUNK);
-    let position = startOffset;
-    let remaining = typeof totalLength === 'number' ? totalLength : Infinity;
-    let bytesSent = 0;
-
-    const readNext = (): void => {
-        if (remaining <= 0) {
-            finish();
-            return;
-        }
-        const want = Math.min(CHUNK, remaining);
-        fsRead(fdArg, buffer, 0, want, position, (err, bytesRead) => {
-            if (err) {
-                cleanup(err);
-                return;
-            }
-            if (bytesRead === 0) {
-                finish();
-                return;
-            }
-            position += bytesRead;
-            bytesSent += bytesRead;
-            remaining -= bytesRead;
-            // Copy the chunk so the same backing buffer can be reused on the
-            // next read iteration without overwriting in-flight Soup data.
-            const slice = Buffer.allocUnsafe(bytesRead);
-            buffer.copy(slice, 0, 0, bytesRead);
-            const ok = res.write(slice);
-            if (ok) {
-                readNext();
-            } else {
-                res.once('drain', readNext);
-            }
-        });
-    };
-
-    const finish = (): void => {
-        res.end();
-        if (closeFd) {
-            try {
-                closeSync(fdArg);
-            } catch {
-                /* ignore */
-            }
-        }
-    };
-
-    const cleanup = (err: Error): void => {
-        if (options.onError) options.onError(err);
-        else res.destroy(err);
-        if (closeFd) {
-            try {
-                closeSync(fdArg);
-            } catch {
-                /* ignore */
-            }
-        }
-    };
-
-    // Suppress empty-body fstat path: if length===0 we just close out.
-    if (remaining === 0) {
-        finish();
-        return;
-    }
-
-    readNext();
-    // Mark that we used the fd-streaming path so listeners know the body
-    // is being delivered out-of-band of the regular write() machinery.
-    void bytesSent;
-}
-
-/**
- * _fdPath — best-effort fd → path lookup via `/proc/self/fd/<fd>`.
- *
- * Used only for statCheck; `fs.statSync` accepts that path on Linux to
- * stat the open FD. Returns null on non-Linux (caller falls back to
- * `/proc/self/fd/N` regardless — `statSync` will fail cleanly).
- */
-function _fdPath(fd: number): string | null {
-    if (typeof fd !== 'number' || fd < 0) return null;
-    return '/proc/self/fd/' + fd;
-}
+installHeaderMethods(Http2ServerResponse.prototype);
+installStreamIoMethods(Http2ServerResponse.prototype);
+installRespondWithFileMethods(Http2ServerResponse.prototype);
+installPushMethods(Http2ServerResponse.prototype, ServerHttp2Stream.prototype);
